+++ src/components/ui/searchbar/searchInput.tsx
@@ -1,16 +1,16 @@
 import React, { FC } from "react";
 import debounce from "@/shared/utils/helpers/debounce";
-import classes from "@/ui/searchbar/searchbar.module.scss";
+import classes from "@/components/ui/searchbar/searchbar.module.scss";
 
-type reactEventProp = (event: React.ChangeEvent<HTMLInputElement>) => void;
+type inputChangeEvent = (event: React.ChangeEvent<HTMLInputElement>) => void;
 
 interface ISearchInputProps {
-  onChange: reactEventProp;
+  onChange: inputChangeEvent;
   loader: (isActive: boolean) => void;
 }
 
 const SearchInput: FC<ISearchInputProps> = ({ onChange, loader }) => {
-  const debouncedOnChange: reactEventProp = debounce(onChange, 1000);
+  const debouncedOnChange: inputChangeEvent = debounce(onChange, 1000);
   const onChangeHandler = (event: React.ChangeEvent<HTMLInputElement>) => {
     loader(true);
     debouncedOnChange(event);